put prism ops in same file
