--- src/math/operations/prismOps/index.ts
@@ -1,3 +1,0 @@
-export { elongate, gyroelongate } from "./elongate"
-export { shorten } from "./shorten"
-export { turn } from "./turn"
